+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 - Handle composed modules when expanding a subworkflow [#6](https://github.com/mirpedrol/nf-class/pull/6)
 - Modify `expand-class` command to avoid conditional includes [#8](https://github.com/mirpedrol/nf-class/pull/8)
 - Don't allow expanding a subworkflow in a pipelines repo & pin nf-core version [#10](https://github.com/mirpedrol/nf-class/pull/10)
+- Remove `prefix` and `suffix` options from `expand-class` command [#11](https://github.com/mirpedrol/nf-class/pull/11)